// src/app/shared/models/Company.models.ts
export class Company {
  id:number;
  name: String;
  primaryConatct:String;
  addres_1: String;
  address_2: String;
  city: String;
  ext:number;
  email:String;
  zip:String;
  tollFree:String;
  state: String;
  country: String;
  currency:String;
  paymentTerms:String;
  telephone: String;
  taxID:String;
  note:String;
  faxNumber: String;
  remittanceEmail: String;
}